Integrate auth and user structure
