+++ src/main/java/life/inha/icemarket/IceMarketApplication.java
@@ -1,13 +1,8 @@
 package life.inha.icemarket;
 
-import life.inha.icemarket.domain.core.User;
-import life.inha.icemarket.respository.UserRepository;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
 
 @SpringBootApplication
 @EnableJpaAuditing
