+++ src/main/java/life/inha/icemarket/controller/ChatController.java
@@ -1,7 +1,7 @@
 package life.inha.icemarket.controller.core;
 
-import life.inha.icemarket.domain.core.Room;
-import life.inha.icemarket.domain.core.User;
+import life.inha.icemarket.domain.Room;
+import life.inha.icemarket.domain.User;
 import life.inha.icemarket.respository.ChatRepository;
 import life.inha.icemarket.respository.RoomRepository;
 import life.inha.icemarket.respository.UserRepository;
