+++ src/main/java/life/inha/icemarket/controller/MyPageController.java
@@ -4,7 +4,7 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import life.inha.icemarket.domain.auth.User;
+import life.inha.icemarket.domain.User;
 import life.inha.icemarket.dto.MyPageDto;
 import life.inha.icemarket.service.auth.MyPageService;
 import lombok.RequiredArgsConstructor;
@@ -12,10 +12,9 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
-import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
 
 @Tag(name = "MyPage", description = "마이페이지")
 @RestController
@@ -32,7 +31,7 @@ public class MyPageController {
     })
     public ResponseEntity<MyPageDto> getUserInfo(@PathVariable Integer userId) throws Exception { // 추후 @Autenticated로 UserId 가져오도록 수정 필요
         HttpHeaders header = new HttpHeaders();
-        header.setContentType(new MediaType("application", "json", Charset.forName("UTF-8")));
+        header.setContentType(new MediaType("application", "json", StandardCharsets.UTF_8));
         return new ResponseEntity<MyPageDto>(myPageService.getMyPageDto(userId), header, HttpStatus.OK);
     }
 
@@ -46,7 +45,7 @@ public class MyPageController {
     })
     public ResponseEntity<User> updateUserInfo(@PathVariable Integer userId, @RequestParam("nickname") String nickname) throws Exception {
         HttpHeaders header = new HttpHeaders();
-        header.setContentType(new MediaType("application", "json", Charset.forName("UTF-8")));
+        header.setContentType(new MediaType("application", "json", StandardCharsets.UTF_8));
         return new ResponseEntity<User>(myPageService.update(userId, nickname), header, HttpStatus.OK);
     }
