+++ src/main/java/life/inha/icemarket/domain/Room.java
@@ -1,4 +1,4 @@
-package life.inha.icemarket.domain.core;
+package life.inha.icemarket.domain;
 
 import lombok.*;
 import org.springframework.data.annotation.CreatedDate;
@@ -7,14 +7,17 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.*;
 import java.time.Instant;
+import java.util.List;
 
+@AllArgsConstructor
 @RequiredArgsConstructor
 @NoArgsConstructor
+@Builder
 @Getter
 @Setter
-@Entity(name = "users")
+@Entity(name = "rooms")
 @EntityListeners(AuditingEntityListener.class)
-public class User {
+public class Room {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
@@ -22,14 +25,9 @@ public class User {
     @NonNull
     private String name;
 
-    @NonNull
-    private String email;
-
-    @NonNull
-    @Column(name = "password_hashed")
-    private String passwordHashed;
-
-    private String nickname;
+    @ManyToMany
+    @JoinTable(name = "room_users", joinColumns = @JoinColumn(name = "room_id"), inverseJoinColumns = @JoinColumn(name = "user_id"))
+    private List<User> users;
 
     @CreatedDate
     private Instant createdAt;
