+++ src/main/java/life/inha/icemarket/dto/MyPageDto.java
@@ -1,14 +1,18 @@
 package life.inha.icemarket.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import life.inha.icemarket.domain.core.Post;
-import lombok.*;
+import life.inha.icemarket.domain.Post;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
 
 import java.util.List;
 
 @Builder
 @AllArgsConstructor
-@Getter @Setter
+@Getter
+@Setter
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class MyPageDto {
     private String name;
