+++ src/main/java/life/inha/icemarket/respository/UserRepository.java
@@ -1,6 +1,6 @@
 package life.inha.icemarket.respository;
 
-import life.inha.icemarket.domain.core.User;
+import life.inha.icemarket.domain.User;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
