+++ src/main/java/life/inha/icemarket/service/MyPageService.java
@@ -1,8 +1,8 @@
 package life.inha.icemarket.service.auth;
 
-import life.inha.icemarket.domain.auth.User;
-import life.inha.icemarket.domain.auth.UserRepository;
+import life.inha.icemarket.domain.User;
 import life.inha.icemarket.dto.MyPageDto;
+import life.inha.icemarket.respository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,7 +14,7 @@ public class MyPageService {
 
     private final UserRepository userRepository;
 
-    public User update(Integer userId, String nickname) throws Exception{
+    public User update(Integer userId, String nickname) throws Exception {
         User user = userRepository.findById(userId).orElseThrow(() -> new Exception("사용자를 찾을 수 없습니다."));
         user.setNickname(nickname);
         return userRepository.save(user);
