+++ src/main/resources/application.properties
@@ -1,11 +1,5 @@
-#spring.jpa.hibernate.ddl-auto=create
-#spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:3306/${MYSQL_DATABASE:icemarket}
-#spring.datasource.username=${MYSQL_USER:ice}
-#spring.datasource.password-file=${MYSQL_PASSWORD_FILE:db/password.txt}
-#spring.jpa.show-sql=true
-
-spring.jpa.hibernate.ddl-auto=create
-spring.datasource.url=jdbc:mysql://localhost:3306/icemarket
-spring.datasource.username=root
-spring.datasource.password=family0831
+spring.jpa.hibernate.ddl-auto=none
+spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:3306/${MYSQL_DATABASE:icemarket}
+spring.datasource.username=${MYSQL_USER:ice}
+spring.datasource.password-file=${MYSQL_PASSWORD_FILE:db/password.txt}
 spring.jpa.show-sql=true
