--- src/test/java/life/inha/icemarket/domain/auth/dummy.txt
@@ -1,2 +1,0 @@
-프로젝트 디렉토리 생성을 위한 Dummy 파일입니다.
-Local에 Pull 받은 후 삭제해주세요.
